refactor(perf): 回应 review——预览链约束回补 + 常量迁 tuning.py + import 对齐
🟡 预览链的 thread_count=1 是浏览器硬解 + zerolatency 硬约束,合进共享常量后
定义处只剩吞吐理由,改值的人会踩坑:
- tuning.py 的 ENCODE_THREADS 补 ⚠ 说明:>1 会退帧级并行、破 zerolatency
- transcoder.py 那句 "threads=1 forces..." 已读不出 1,改指向 ENCODE_THREADS

🔵 常量迁 miot/tuning.py(不含实现):ENCODE_THREADS 在 miot 内零引用、三处消费全
在 miloco;此前 miloco 为拿一个 int 把 decoder.py(连带 PIL/subprocess)import 进
感知热路径。新模块只放调参常量,跨包面回到接口/配置类模块。

🔵 speech_vad 的 TINY_MODEL_THREADS 与 dedup_embedder 对齐走模块级 import(整数无
惰性约束),惰性注释收窄到只解释 apply_kleidiai_opt_out helper。

🔵 抓片段测试 finally 补关 mp4 容器(先容器后线程池),不留半截清理的示范。
