+++ backend/miot/src/miot/decoder.py
@@ -24,18 +24,10 @@ from av.video.frame import VideoFrame
 from PIL import Image
 
 from .error import MIoTMediaDecoderError
+from .tuning import DECODE_THREADS, SWSCALE_THREADS
 from .types import MIoTCameraCodec, MIoTCameraFrameData
 
 _LOGGER = logging.getLogger(__name__)
-
-# 单路实时解码/像素转换/串行编码都远用不满满核线程池,统一钉死上限。
-# 想知道某个值用在哪,grep 常量名即可——不在此维护站点清单(清单会过期)。
-DECODE_THREADS = 4   # FFmpeg 解码器 slice 线程,含余量
-SWSCALE_THREADS = 2  # 独立 swscale 像素转换:1 已够,留 1 压单帧尖峰
-ENCODE_THREADS = 1   # libx264:每条 BGR→H.264 链各自已串行跑在专属线程里,内部并行
-                     # 买不到吞吐。设它也顺带把 encode 内部那次 bgr24→yuv420p 的
-                     # swscale 收敛到 1——PyAV 取的就是 codec.thread_count
-                     # (av/video/codeccontext.py:95)。
 
 
 # H.264 NAL unit types that contain an IDR coded slice.
